+++ backend/src/main/java/com/hjjang/backend/global/config/security/properties/AuthProperties.java
@@ -1,13 +1,12 @@
-package com.hjjang.backend.global.config.properties;
+package com.hjjang.backend.global.config.security.properties;
 
-import java.util.ArrayList;
-import java.util.List;
-
+import lombok.Data;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 
-import lombok.Data;
+import java.util.ArrayList;
+import java.util.List;
 
 @Data
 @Configuration
